LAB5/todoviewset/requirements.txt
asgiref==3.3.1
cffi==1.14.5
cryptography==3.4.6
Django==3.1.7
django-rest-framework==0.1.0
djangorestframework==3.11.2
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
